We're secretly replacing the terrain in WRF's nest with new Terrain Blend. Let's watch.
But seriously...

JM This is the fix from early this morning, plus a modification to the
JM blend_terrain subroutine (dyn_em/nest_init_utils.F) that forces the code to
JM use only interpolated terrain from the parent over the entire spec_bdy_width
JM and then blends the terrain over a zone of five additional cells into the
JM interior. It looks like this:
JM
JM i=  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17
JM
JM     @  *  *  *  *  *  *  *  *  *  -  -  -  -  -  o  o
JM
JM   @ is the outermost cell on the boundary and entirely specified by parent
JM
JM   * is the relax_zone of the boundary region
JM
JM   cells marked @ and * together make up the spec_bdy_width (10) for the
JM   domain the terrain for these cells is interpolated from the parent domain.
JM
JM   - is the "blend zone" where the interpolated terrain is gradually replaced
JM     with the high-res read-in terrain as we move inward. In the current run,
JM     cell 11 is 1/7 high res terrain, cell 12 is 2/7, cell 3 is 3/7, and so
JM     on until we get to cell 16, the first cell of the "interior" (o), where the
JM     terrain is 7/7 or 100% high-res.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1433 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d