Two changes to the WRF code with this commit.
1) Update the Registry.EM_CHEM registry file so that the defaults
match-up with the Registry.EM file.  No impact to the simulations should
occur as the user specifies these values in the namelist.input file.

2) Modify the calculation of RMOL in phys/module_sf_sfclay.F   There are
some issues with the chemical deposition based upon the current release.
This is due to RMOL being off by as much as three orders of magnitude.
  The updated code changes the calculation of RMOL so that more
reasonable values are obtained.

The variable rmol is not being used elsewhere in the phys, frame, share,
dyn_em directories (it is initialized to zero in
dyn_nmm/module_physics_calls.F) so the meteorology only WRF results are
not impacted.  There should be an impact on the chem baseline as
chemical deposition is increased.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1816 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d