First phase of changes to dereferene grid% inside mediation layer (solve_em) rather than in long argument list. JM
Additional information (from email to wrf developers list, 8/24/2006):

The problem and the fix
-----------------------

A while ago -- early in the design of the WRF software -- we ran into a
problem with Fortran 90. Specifically, we found we couldn't use the fields
in the "grid" derived data type directly in the code without severe
performance penalties. So, as a workaround, we moved the dereferencing
(selection of fields out of the "grid%" container) higher up in the call
tree, above the call to the solver, and passed all the individual state
fields through the argument list to solve_em. That worked until recently
when, with the addition of new packages and state variables, the argument
list finally grew too long. Meantime, however, many of the compilers got
smarter and the penalty for dereferencing the fields from grid% within solve
has mostly gone away. This means we can go back closer to the original
design: just pass "grid" through the argument list to solve_em and
dereference the fields where they're used.  So, for example, a reference to
"u_2" in the current (repository) solve_em will now appear as grid%em_u_2
(we'll get rid of that "em_" prefix too, eventually).

Earlier this week, I recoded a test version of solve_em to work this way and
ran some tests on Linux and on Bluevista. For the most part, it works as
hoped, and the performance looks good comparing timings for sections of the
solver (see attached graphic -- the blue is old; lavender is new). There is
one complication, involving 4d arrays. Although the compilers now handle
passing grid%em_u_2 correctly without generating performance-degrading and
memory-wasting copies, they do not seem to quite understand yet that they
can do the same with 3-D elements of 4D arrays. Second attachment is my note
to IBM this morning, providing more detail on that and requesting action on
their part. But our situation is that even if IBM agrees their compiler is
doing the wrong thing, it's not going to result in any immediate resolution
of the problem until the next release of their compiler. And we have to stay
backward compatible with older versions of compilers. Therefore, the "fix"
I've come up with is a hybrid.

The fix I'm proposing will remove *most* state arrays from the argument list
but the 4d arrays and their related 4d boundary arrays will continue to be
dereferenced above the call to solve_em and passed through the argument
list. The rest of the state arrays, which don't cause performance problems,
will be dereferenced in-place within the solve_em routine itself. This will
have the advantage of drastically shortening the argument list to solve_em
to just a dozen or so arguments, including grid and config_flags, and will
effectively eliminate the obstacle we've run into with the infrastructure.
It also eliminates the need for some other non-intuitive aspects of the
solver: the includes of em_scalar_derefs.inc at the beginning and end of the
solve routine and the mysterious business with deref_kludge. The
disappointing aspect of this is that we still can't completely eliminate
passing individual state arrays through the argument list or the need for
the Registry to generate these lists. But it's probably the best we can do
at this time.

I've placed two versions of solve_em in ~michalak/SOLVER for you to inspect
and compare. The file solve_em.F.current is what we have now in the
repository. The file solve_em.F.new is the proposed changed one. Solve_em
and the call to solve_em from solve_interface are the only affected
routines. Model layer subroutines, physics interfaces, etc. and not affected
at all.





git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1926 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d