MAJOR CHANGES TO GD CONVECTIVE PARAMETERIZATION
(1) Introduction of a parameter "closure(n)", to keep track of how many 
closures are allowed to contribute to the precipitation. If a closure is  forced out of the picture at all times, this is considered when 
calculating the final mass flux (e.g., if the mass flux is determined by  4 closure, not 16, the summed up mass flux is divided by 4, not 16). This is a significant change that will tend to increase the maximum values of non-resolved precipitation in some areas (probably biggest effect is over water). 


(2) GSW and xland are used inside the scheme to shut off the possibility of convection under certain conditions. Most noticeable effect is to turn down convection at night when large caps are present. 


(3) GD_CLOUD and GD_CLOUD2 are introduced. These are variables that hold the in-cloud hydrometeor mixing ratios (as modeled by the convective parameterization). These are only used in WRF-Chem runs, in which they are coupled directly to the radiation (atmospheric and UV). 


(4) fix to some array problem that existed, also other minor clean-up stuff....


ARW/NMM PHYSIC PACKAGE CHANGES

ARW physics options can now be used in the NMM core simulations

The RUC LSM has some minor bug fixes.  In addition, we have added the capability to factor in frozen precipitation into the calculations.

Added additional flexibility with the number of soil levels.  Was previously set to 6 levels only.  Now the user can specify more levels (six is the minimum number).

Added the capability to initialize hydrometeors, surface and soil fields in the NMM core from the RUC model

And there is a new version of the Thompson microphysics scheme which will work for both dynamical cores.


CHEMISTRY PACKAGE CHANGES	


(1) Adding KPP to the chemistry options. The Kinetic Pre-Processor automatically generates the chemical mechanisms for use in WRF-Chem. This is a widely used tool in the atmospheric chemistry community, easy to use, and it makes it much simpler to modify the otherwise very complicated chemical mechanisms.  Added environmental variable WRF_KPP to compile this section of the code  default is zero (0) which means that the KPP code is not compiled.

(2) Generalization of the chem_driver and chem_init routines to accommodate the NMM core. 

(3) Bug fixes and updates throughout the chemistry modules. 

(4) Added advection of chemical species in the NMM core.

(5) The chemistry code now allows 1-way nesting through modification of ndown and introducing a new boundary condition option in module_input_chem_data.F.

(6) Have added namelist variables to control the chemistry boundary conditions (have_bcs_chem) as well as the convection/radiation feedback (cu_rad_feedback)

(7) Added features to allow compilation as well as cleaning of the new (and old) chemistry code.

--------------------------------------------------
Georg Grell, Steven Peckham, Tanya Smirnova and the rest of the WRF gang.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@1980 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d