1) Fix external/io_netcdf/makefile so that diffwrf compiles; this entails adding an include path to a cpp line.
2) Additional changes for WRF_FLOAT need to be
applied to gen_wrf_io.c. There is some code in
gen_wrf_io.c that was added since the original
WRF_FLOAT changes to the file, and this new
code requires the WRF_FLOAT changes, too.

3) Add #ifdef MAC_KLUDGE logic to module_physics_init.F and
module_ra_cam.F to allow WRF to compile with xlf on Mac systems;
however, not all of the CAM radiation scheme can be compiled
under this configuration, so a fatal message is also added to
stop when this is used with Mac/xlf.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2014 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d