New feature: allow WPS for NMM
The auxinput1_inname line in the registry.io_boilerplate
file was removed and placed in two separate boilerplate
files so that the core specific portion of the name is
defined correctly (the em file is a default for the em
real program).  Each of the EM and NMM Registry files
has an additional include (the new 1-line that was
removed from the original boilerplate io file).

No changes to results.





git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2190 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d