Bug Fix - should be posted as a known problem, will change results
The observation nudging namelist has two
errors.
1) The interval and ending times for the
obs nudging are for each domain, and only
a single entry is given.  The impact is that
obs nudging only occurs on the coarse grid.
2) The time window units are hours, not minutes.
The default nml value of 40 means that obs are
used for nearly 80 hours.

The mention of the units (h) is placed in the
README.namelsit file.  Accompanying that mod
is a comment about the diffusion defaults
that are suggested for real-data cases.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2192 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d