enhancement ----------- Does not change results.  If user is supplying SI as the input, AND the user has more than 100 time periods to process, the code now works.
Description
-----------
The extension (such as ".nc") on the WPS data and 
the non-extension on the SI data are both assumed
by default.  This removes the need to try to 
initially open an SI file that is missing the extension,
and then do a re-open without the extension.

Impacted files
--------------
M    real_em.F
M    real_nmm.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2217 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d