type of change: --------------- enhancement Stop horizontally interpolating the middle point value for lat/lon for domains that have a valid value.  This changes the metadata, and will impact Var code in a cycling situation where these map proj fields are used to determine whether an observation is used.  This impacts ARW only.
purpose of change:
------------------
For model nests that have no input file, the metadata is
derived from the parent domain.  The "middle" lat and lon
of the domains has to be interpolated.  For domains with
an input file, this is not necessary.  The better solution
(and one that is coming), is to have the point's lat/lon
directly computed.

Impacted files:
---------------
M    start_em.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2233 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d