5 minor changes  - mostly related to chem code modifications.
1) Bug fix in module_radiation_driver.F 
move the cu_rad_feedback if block at the end of
module_radiation_driver to an earlier location in the routine. This
prevents an undefined variable.

2) minor addition to dyn_em/start_em.F and chem/chemics_init.F
Initialize gd_cloud and gd_cloud2 arrays. While this should not cause
any problems, it is better to give this ounce of prevention. This was 
left out of the earlier commit.

3) John Michalakes pointed out that we had an issue in Registry.EM_CHEM
that generated a warning from the registry program. One line removed
from Registry.EM_CHEM and Registry.NMM_CHEM as registry.io_boilerplate_EM
 is not in thei repository code.

4) bug fix in phys/module_mixactive.F

swap 2 lines 

integer :: k, kmn, kms
real  dimension(..., kmn, kms,...) :: chem

This is so that the the integers used in chem are declared before the
chem array. 

5) Bug fix in phys/module_ra_gsfcsw.F

Set the single scattering albedo max to 1.0 instead of 0.9



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2276 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d