Small bug fixes plus assorted updates of NMM codes to reflect changes    made in production during the past 6 months.
1) Bug fix in dyn_nmm/module_initialize_real.F.  There are lines where values
for soil depth and root depth are hardwired.  These lines were previously
commented out so as not to override the input but apparently in the code
cleanup that was committed on 1/19/2007 these lines were uncommented.  They
are now being commented out again. 
2) Bug fixes in phys/module_ra_gfdleta.F.  The value of DEGRAD taken from
share/module_model_constants.F is being misused by taking its inverse for
conversion from radians to degrees of latitude.  This affects only ozone
computations.  Also there are local declarations of DEGRAD that disagree with
that in module_model_constants so these local names are now changed.  The
time going into the mean zenith angle computation was slightly off and is
being corrected.  Finally force the working arrays for cloud water/ice to
be non-negative.
3) Bug fix in phys/module_physics_init.F.  Add an IF test for NMM_CORE so that
variable 'usebgalb' is .TRUE. which leads to use of the 2-D input field albedo
instead of table values from LANDUSE_INIT table.
4) Bug fix in dyn_nmm/module_PHYSICS_CALLS.F.  Update an auxiliary convective
precip array that is used only by radiation.
5) Numerous refinements in phys/module_cu_bmj.F.
6) Refinement in phys/module_mp_etanew.F for parameters and logic involving
ice.
7) Refinement in phys/module_sf_lsm_nmm.F.  Final general cleanup prior to
upcoming transition to unified NOAH code.
8) Refinement in dyn_nmm/start_domain_nmm.F.  Change initial values placed
into soil temperature from NMM_TSK to 273.16.
9) Refinement in dyn_nmm/solve_nmm.F.  Replace IF test on whether to call
convection with one based on the selected scheme's numerical value being
nonzero or not.
10) Refinement in dyn_nmm/module_IGWAVE_ADJUST.F.  Enhance the divergence
damping for the external mode.
11) Refinement in dyn_nmm/module_DIFFUSION_NMM.F.  Horizontal diffusion
currently not allowed to act over much of the elevated terrain.  Allow it to
operate in on additional points with steeper slopes.
12) Refinement in dyn_nmm/module_BNDRY_COND.F.  Remove extraneous comments.
13) Reduce history output in Registry.NMM, Registry.NMM_NEST and
Registry.NMM_CHEM.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2292 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d