New Feature: subgrid arrays (scheme 1)
This is a commit requested by the wildfire application people, 
but it might be more generally useful to others. The changes 
allow a user to specify a subgrid array as state array within 
a given domain or domains.  

To specify a subgrid array in the Registry, one adds * to the 
dimension one wishes to have refined. The refinement is relative 
to the grid on that domain.  For example:

  state real firevar *ik*j misc ...

will define a variable named grid%firevar( : , : , : ) where the 
first and third (i and j) dimensions are refined by an integer 
refinement specified by the new namelist variables:

  rconfig integer sr_x namelist.domains max_domains 0
  rconfig integer sr_y namelist.domains max_domains 0

which refine the x (i) and y (j) dimensions of the variable by the 
respective refinment ratios. The sr_x and sr_y variables are per-domain, 
meaning different domains can define different refinements (or none, 
for sr_x=sr_y=0, the default, which means the subgrid array variable 
is not allocated on that domain -- so don't try to use it :-).

Example:

  In Registry:
    state real firevar *ik*j misc 1 - i3h

  In namelist:
    &domains
    sr_x = 0,30,0
    sr_y = 0,40,0

This will define a state array grid%firevar(:,:,:) and, on domain 2, 
allocate it as:

  ALLOCATE(grid%firevar((sm31-1)*sr_x+1:em31*sr_x,sm32:em32,(sm33-1)*sr_y+1:em33*sr_y),STAT=ierr)

and for each cell in the principal grid of the domain, the firevar 
variable will have 30x40 (1200). There's no upper limit on refinement 
except available memory.  I've heard tell of 1000x1000 subgrids. One 
could also think of refining only one dimension and using the subgrid 
for ensemble members.

Note that if sr_x and sr_y are zero, the first and third dimensions go 
from 1:0 which, I think, will actually allocate a few bytes to the data 
structure (that must be accessed backwards!) but negligable.

There is a new subroutine added to module_domain for getting the subgrid
dimensions from the WRF infrastructure:

  CALL get_ijk_from_subgrid ( grid , &
                            sids, side, sjds, sjde, skds, skde, &
                            sims, sime, sjms, sjme, skms, skme, &
                            sips, sipe, sjps, sjpe, skps, skpe )

and for convenience, declarations of these new variables and a call to 
this routine are included in the ARW solver, dyn_em/solve_em.F. 
The translation from a grid index to a subgrid index is:

    si = (i-1)*sr_x+1 and sj = (j-1)*sr_y+1

the inverse (from subgrid to grid) is:

    i = (si-1)/sr_x + 1 and j = (sj-1)/sr_y + 1

I/O and HALO exchanges work on these subgrid variables. Nesting will 
not, since the WRF software framework supports interpolation between 
domains, not within a domain. It is assumed that if a user calls for 
a subgrid variable they will provide the mechanism for interpolating 
data between the grid and subgrid arrays on a domain.

The k dimension can not be subgridded -- however, it's already 
possible to define different vertical extents for 3d variables using 
dimspec entries in the registry.

The changes will have no affect on any WRF application unless they 
are used (specified in the registry and activated in the namelist at 
run time).

Affected files for this commit:


M    external/RSL_LITE/gen_comms.c
M    frame/module_domain.F
M    tools/gen_wrf_io.c
M    tools/type.c
M    tools/data.h
M    tools/gen_interp.c
M    tools/misc.c
M    share/mediation_feedback_domain.F
M    Registry/registry.io_boilerplate
M    dyn_em/solve_em.F


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@2298 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d