JM WRF-1. Adding TRADFLAG to builds of diffwrf
This was reported by a number of users on a number of systems (usually some
Linux variant) once V3 went out.  Diffwrf is a utility, so this was not a fatal
error but some users (esp. vendors) do need it. This fix makes it compile
correctly when CPP needs -traditional.

Affected file: Makefile (top-level)

WRF-2. Fix transpose routine collisions between Grib[12] and PHDF5 (from Bob
Numrich)

Numrich tried (and, I think, ultimately gave up) to get PHDF5 to work on a Linux
system (I think it was an SGI named Calhoun?) but in the course of this he
discovered that the transpose routines in the grib implementations of the I/O
API collide with the hdf5 transpose (neither set is contained in a module,
unfortunately).  This isn't an issue as long as one doesn't try to use PHDF5,
but we do claim to support it. This fix simply renames the transpose routines to
transpose_grib and transpose_hdf5.

Affected files: 
   external/io_phdf5/wrf-phdf5.F90
   external/io_grib_share/io_grib_share.F
   external/io_grib1/io_grib1.F
   external/io_grib2/io_grib2.F

WRF-3.  Rules for building netcdf and pnetcdf I/O APIs should use $(NETCDFPATH)
and $(PNETCDFPATH), not CONFIGURE_NETCDF_PATH or CONFIGURE_PNETCDF_PATH.  Sort
of cosmetic, because it works the way it is.

Affected file:
   arch/postamble_new

WRF_4.  Fix the problem with very long argument lists that are non-standard
Fortran and causing problems for certain older compilers. The fix has been added
to the tools/standard.c program, which currently makes sure that WRF_ERROR_FATAL
expansions don't exceed 132 cols. Under the new capability, standard.c (which is
invoked as part of the build for every .F file) looks for call statements to
surface_driver, radiation_driver, cumulus_driver, and pbl_driver that appear in
dyn_em/module_first_rk_step_prep_part1.F (ARW) and
dyn_nmm/module_PHYSICS_CALLS.F (NMM). It then removes from the argument list out
all the white space, comments, and mashes it all up into an ugly but
space-saving and more importantly Fortran-standard block of continued statements
in the .f90 file that the compiler sees. It also takes care not to mess up the
#ifdef/#endif pairs that appear in some of these lists.

Question for NMM developers: since this is a little black-boxy, would you rather
that this wasn't done to the NMM argument lists in module_PHYSICS_CALLS.F?  It's
easy to put a switch into the standard.c program that would turn off this
translation if it sees it's compiling for NMM.

Affected file:
   tools/standard.c



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3085 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d