This is a fix for a problem with nesting using non-3 parent:nest refinement ratios, one of the issues on the V3.1 checklist. Problem occurred with both moving and non-moving nests. Fix involves correcting a number of instances where a hard-coded nest ratio of 3 was assumed, either directly or in the specification of halo-widths associated with shifting arrays when moving domains.
The affected files are:

  external/RSL_LITE/gen_comms.c
  share/interp_fcn.F
  Registry/Registry.EM

The fix to the Registry file simply removes an out-of-date halo specification that I noticed while looking into this problem.

Wei has tested the proposed fixes for nesting ratio of 5 and determined that the solution artifacts she had been seeing (processor boundaries) are now gone and the solution is clean.
            



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3543 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d