I am proposing a set of commits to finish up this item in the Release committee's release "picture".
>  1.b) 5-6 dimensional tracer capability
>       Note: Communications/nesting bugfixes to be completed

Specifically, this finishes up the set of commits that were first introduced
to the repository on 12/1/09 as: 

   "Change to allow extra dimensions to 4D scalar arrays.  This is to 
    satisfy a request from some users - e.g. some chemistry folks -- to 
    let scalar arrays have more dimensions than just 4."

The capability was introduced with I/O supported, but halo exchanges and the
ability to use these arrays with nesting was not included at that time.
This commit adds those capabilities too. The 5 or greater dimensional tracer
arrays can now be included in HALO declarations and take part in nesting,
including moving nests.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3546 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d