These was item 1b in my 2/13/2009 proposal -- mods to dyn_em/module_advect_em.F and dyn_em/module_big_step_utilities.F for overdecomposition.  The problem occurred compiling and running OpenMP on the IBM bluefire.ucar.edu.  Every OpenMP test failed.  As a result, I withdrew these, and committed and tagged the remainder of my changes.
Subsequently, I found a workaround that involves a modification to the
WRF build. This update puts back the original 1b mods plus updates to the
three files in the arch directory: postamble_new, noopt_exceptions, and
noopt_exceptions_f.  These adds a test to the compilation that only
invokes OpenMP if the source file contains OMP directives.

Removed a hack in regtest.csh script that adds :noopt to -qsmp=noauto.
This was originally in there to achieve bit-for-bit for OpenMP runs
on AIX.  Not needed any more, with change above.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3620 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d