Bug found by Al Bourgeois in the obs nudging code:
I found a bug that can result in tons of printout to the rsl standard and error files.

Starting on line 1807, there is the section of code:

! jc arbitrary check here
            IF (iprt) THEN
              if (kpblt(i).gt.25 .and. ktau.ne.0) &
                                     write(msg,552)inest,i,j,kpblt(i)
552 FORMAT('kpblt is gt 25, inest,i,j,kpblt=',4i4)
              call wrf_message(msg)
            ENDIF


As you can see, the "if (kpblt" conditional does not have a "then" associated, so the "call wrf_message" just below it is not with the conditional. So the wrf_message gets called all the time, which produces tons of output. The fix is to put the call to wrf_message within the conditional:

! jc arbitrary check here
            IF (iprt) THEN
              if (kpblt(i).gt.25 .and. ktau.ne.0) then
                 write(msg,552)inest,i,j,kpblt(i)
                 call wrf_message(msg)
              endif
552 FORMAT('kpblt is gt 25, inest,i,j,kpblt=',4i4)
            ENDIF

M    phys/module_fddaobs_rtfdda.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3780 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d