Changes to the top-four big-spender routines to reduce the static memory usage of WRF (see 6/8/09 email subj: top memory users in WRF). Three of the four changes involve changing the way lookup tables are stored in the CAM radiation and 2 Thompson MP schemes. Currently, these lookup tables are defined as module data; now, the tables are instead defined as ALLOCATABLE and then allocated during physics initialization and only if the routines have been selected at run time. The two Thompson schemes have also been modified so that the tables will only be initialized once, not every time a new domain is instantiated.
The fourth change reduces the number of 30KB string buffers in the
GRIB1 implementation of the WRF I/O API from 400 down to 60. The
buffers are used to store and output time dependent and time
independent metadata. There are two such buffers for each of 200
allowed for datasets. Given that WRF 3.1 doesn't have more than a
couple of dozen I/O streams, it's impossible for it to use that many
open datasets, even if every stream were reading or writing through
GRIB1.

Prior to the changes the UNIX size command reports a static size of
130MB for wrf.exe. This is the bedrock per-process memory usage and
does not include dynamically allocated and stack data once the program
starts to run. With the changes the static size is 32MB -- much closer
to the old V2.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3868 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d