1.  (original 6/17) Proposal to fix G3 cumulus so that it works with OpenMP.
Gets rid of the regtest FAILs.  Two files affected:
 
   phys/module_cu_g3.F
   phys/module_cumulus_driver.F
 
The fix involved modifying a couple places in the spread3d routine that control whether to compute out on tile halos or not; also made local variables tile dimensioned (sometimes plus a halo) instead of patch dimensioned; and pushed the OpenMP directives up into the driver.  Georg has reviewed and approved.
 
Diffs file attached, and the full routines are on bluefire in /ptmp/michalak/CU5/phys .
 
2.  (added 6/18) Proposal to fix OpenMP bug in surface driver (reported by Jimy)
 
This is to fix an issue with the value of the myj flag coming out of the first OpenMP loop. The value of the flag was mysteriously changing and causing problems for surface routines in the second OpenMP loop.  The problem was that that myj, a shared variable, was being assigned multiple times by each thread inside the first OpenMP loop. Nothwithstanding the fact that the value assigned is always true, and that the assignments were not even executed, OpenMP has to do something special in such cases to make sure that thread-local copies of the myj variable inside the OpenMP loop get reconciled back to the shared copy of the logical variable at the end of the OpenMP loop. 
 
And that something special, in this case, was wrong.  However, it is not clear that this is a bug in IBMâs compiler because the code itself is probably non-conforming.  The code as written could generate a race condition if at run time each thread took a different path through the SELECT statement.  They donât, but itâs difficult for the compiler to determine that at compile time.  In general, OpenMP loops that have each thread setting the same shared memory location need some kind of extra logic or directives to resolve the potential race condition â or just avoid doing it.
 
The propoosed solution moves the conditional assignment of the myj flag to before the first OpenMP loop.
 
(old is < , new is >)
diff -iw /ptmp/michalak/CU5/phys/module_surface_driver.F phys/module_surface_driver.F
834c834
<   myj = .false.
---
>   myj = ( (sf_sfclay_physics .EQ. MYJSFCSCHEME) .OR. (sf_sfclay_physics .EQ. QNSESFCSCHEME) )
997d996
<         myj =.true.
 
Only affected file is module_surface_driver.F.  In addition to the above differences, a couple of changes to indentation/capitalization.
 
3. Do the same for the isisfc variable: also module_surface_driver.F

4. set SNOBS for non-nudged runs (from JD): Affected file: module_sf_pxlsm.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3876 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d