Two changes to be added to the chemistry code.  The impacted routines are:
1) chem/convert_emiss.F

The emissions conversion program needs to have the use statements updated to fix compile issues on the IBM. There is no impact on any of the WRF model runs and this routine is not tested in the regtest runs.

2) chem/module_add_emiss_burn.F

Additional package names need to be added to a case switch so that biomass burning will be included with additional chemical mechanism options.




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3898 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d