Two quick commits to get into the repository before version 3.1.1
1) Add a simple landuse check and  error message if USGS landuse
was not used with chemistry (deposition does not work wtih other
landuse options at this time).  This change impacts chemistry only.

2) A problem with running DFI with chemistry turned on was brought 
to my attention.  The solution required a simple addition of code 
to the share/dfi.F routine that turns off chemistry for backward 
DFI integration and turns it back on for forward DFI integration.
The change has no impact on non-chemistry runs.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3917 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d