1) Complete set of changes for overdecomposition
These are follow ups to the commit that was proposed and withdrawn 9/4/09.  They allow running on very large numbers of processors in dm+sm mode (4K MPI tasks, each with 4 threads for total of 16K cores). Fixes problems that showed up on a large Blue Gene/P system at Argonne. One set of changes involves a fix to the tiling algorithms in frame/module_tiles.F so that patches that don't have enough cells in a dimension for the number of tiles being requested will still be tiled correctly.

The other set of changes involves fixing loop range calculations in a number of places in the code that were not correct for very fine decompositions of the boundary.  Generally, these original sets of calculations made assumptions about the boundary not being decomposed over tiles.  

That is, always use the min/max functions when setting these.  The incorrect forms of this calculation appear in a number of places in the ARW code and they've been there for a long time.  I never caught them either.  In future, however, always use the second form if adding these kinds of boundary avoidance calculations for loop ranges over a tile computation.

This was a real fun one since it required debugging at scale.  I hope ANL doesn't send me a bill for all the Blue Gene CPU hours I used for debugging runs over the last month.

2)  Modify the local array rfield in the relax_bdy_dry subroutine (dyn_em/module_bc_em.F) to be tile dimensioned rather than memory dimensioned.

This is in keeping with general WRF coding rules for local arrays.  Here, rfield is used as a temporary array for coupling certain state fields with MU before calling relax_bdytend (frame/module_bc.F).

The changed also necessitated creating a separate wrapper for the relax_bdytend routine so that it can handle either a tile-dimensioned field argument (when called for mu-coupled PH, T, and W) or when called for the memory-dimensioned field itself (U, V, or MU).

(3) Add a conditional in share/mediation_nest_move.F so that only process 0 writes information about nest movment.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3958 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d