1) Mods from Portland Group for Windows and a fix
PGI sent a batch of mods that are needed for Windows 32 bit port.  A lot
of them are just good stuff to have anyway -- declaring functions in
C routines and defining prototypes -- and they get rid of some warning
messages we've been seeing on other platforms.  In the case of defining
prototypes, the PGI mods include a definition of a registry routine that
we found and fixed independently because it was causing seg faults for
users running the new MacOS Snow Leopard. So this fix is there too.

2) Remove namelist variables and timing calls that involve months and
time intervals.  This caused a problem for a user who inadvertantly
specified months for one of the namelist variables that control I/O.
Unlike year, day, hour, minute, and second, month is not a regular unit
of time and can't be used to specify intervals (it can only be used to
specify instants in time, and that use is preserved).

3) ESMF interface fixes (reported by James Geiger at NASA GSFC)

Get MPI communicator from ESMF VM.  This code was missing in the WRF-LIS
coupled code but somehow worked anyway on the IBM. Error turned up on
Intel/Linux, which is more careful about skinzers.  Also, put a PRESENT
test around the use of the Sd optional argument in WRFU_TimeGet.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3993 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d