1) This is phase-I of the new I/O capability to allow an arbitrarily large number of input and output streams and to support run-time specification of I/O streams and variable-sets. This phase one commit is just the enabling infrastructure for these capabilities. The code should work exactly as before. You won't notice any more streams or the run-time capability right off the bat, but this will allow those to be added soon.  The change should also lighten the burden on certain compilers by reducing the size of or eliminating registry generated include files associated with I/O.  This will be at the cost of a considerably larger inc/allocs.inc file, which sets up the new run-time data structures for managing the I/O at run-time, and some modest additional memory use for the data-structures themselves. How state variables are organized in model for I/O. A linked-list of nodes containing pointers to all the state variables in WRF along with each variables dimensionality, decomposition, and other attributes is The nodes in the list are TYPE(fieldlist) and are defined in frame/module_domain_type.F.  Each grid datastructure now includes pointers to the head and tail of the list of state variables:
      TYPE ( fieldlist ), POINTER :: head_statevars
      TYPE ( fieldlist ), POINTER :: tail_statevars

The nodes are added to the list as each state variable is allocated in the registry-generated file inc/allocs.inc.  The list is traversed starting at the head with code that looks like this:

      TYPE( fieldlist ), POINTER :: p
        ...
      p => grid%head_statevars%next
      DO WHILE ( ASSOCIATED( p ) )
        IF (on_stream(p%streams,switch)) THEN
          ... do stuff to p%VarName
        ENDIF
        p => p%next
      ENDDO

You'll find two big loops like this in share/input_wrf.F and share/output_wrf.F in place of the numerous includes of registry generated files containing calls to wrf_ext_read_field and wrf_ext_write_field.  The linked lists comprise a node for each field in the order the state entries for those fields appear in the Registry.  This is to make sure that fields are input and output in the order as they were in previous versions of WRF (for applications and I/O formats that are order-dependent).

The logical function on_stream(), defined in frame/module_io.F, takes as arguments a bitmask array p%streams and a switch value defined in the new frame/module_streams.F (e.g. history_only, input_only, aux_hist9_only, ... ) and returns true if the bit corresponding to the stream is set in the bitmask.  By carrying a bitmask around with each variable, it will soon be possible to add or delete variables from a stream at run time.

The p%streams array is arbitrarily long but there will be an equal number of history/auxhist and input/auxinput streams; thus, the total length of the bitmask is 2*the number of history/auxhist streams.  The number of history streams is defined at compile time with the setting of MAX_HISTORY in arch/preamble_new.  At this time it is set to 12, the current number of streams in previous versions of WRF. There are a few other things that need to be done -- increase the number of namelist variables and timekeeping alarms -- before this can be raised.  The registry semantics have been extended so that a two digit stream number can be specified for a variable by using curly braces. For example:

   state real foo  ikj  misc  1  -  i01{12}

will add foo to the main input stream plus auxiliary input streams 1 and 12.

Restart and boundary are still their own streams and are outside the p%streams bitmask. Necessarily, there are no longer specially-named aux_input or aux_hist streams as was done, for example, for the FDDA streams.
Included in these changes:
  - The number of history and input streams is controlled by the value of MAX_HISTORY in configure.wrf (arch/preamble_new).  The total number of streams is 2*MAX_HISTORY, half of which are history (1 main history stream and the rest auxiliary histories) and the other half are input (1 main input stream and the rest auxiliary inputs). That is, if MAX_HISTORY is 35, then there are 35 output and 35 input or 70 total.  The naming of associated time_control namelist variables are the same as before.  There is still only 1 restart and 1 boundary-input stream.
  - One difference with this new scheme is that the streams are all on a per-grid basis, at least internally.  Currently, there's no way to make use of that. The registry specifications of the variable sets associated with a stream will apply to all grids. But once run-time specification is enabled, it will be possible to output different sets of variables on the same stream, depending on the grid.  Frames-per-stream settings still apply singly to all grids.

  - You'll notice that registry.io_boilerplate is a lot shorter.  All the variables that were listed there for controlling streams are now generated automatically based on the number of streams specified in the configure.wrf file (see previous).  - Removed a bunch of registry variables relating to time-series output from out of the restart streams.  It doesn't appear this worked properly in the first place and it was causing problems for the new I/O schemes.
  - Changed FDDA variables from the 'g' stream to '{10}'.
  - Note that the namelist variables that control 'g' and 'sg' fdda in registry.io_boilerplate *still do work* and are still listed in the fdda section of the namelist.input file.  But these associations of special namelist variables with standard time_control variables as well as other special behaviors of namelist variables (e.g. #ifdef PLANET) are now managed in a file dyn_em/namelist_remappings_em.h, which is included by frame/module_configure.F.  Have a look at the contents to see how this is done. Other cores/options can have such files too if needed.

2) Fix for module_esmf_extensions.F90 to only check for existence of Sd variable before computing fractional seconds in a time value.  Also some accumulated fixes for the WRF ESMF module.

3) Remove TOYVAR from NMM registry files.

4) Modify configure script to check for and use timex command for compiles if avail.

5) Have parallel code work when only MPI-1 is available:

  - Modify configure script to add -DMPI2_SUPPORT to ARCH_LOCAL if MPI2 
    is supported
  - Modify wrf_dm_gatherv in external/RSL_LITE/module_dm.F so it works 
    when only MPI-1 is avail (don't use MPI_IN_PLACE arg with MPI_AllGatherv)

6) Changes to stop compilers from whining.  These changes involve modifications to Registry program, Makefiles, and USE statements in various spots throughout program.

  - Split up frame/module_comm_dm.F into module_comm_dm_[0-3].F
  - Split up frame/nl_get* into frame/nl_[sg]et_[0-7]_routines.F
  - Split up frame/module_alloc_space.F into module_alloc_space_[0-9].F

7) Frame directory uses -j for parallel make when available.  Straightened out dependencies in frame/Makefile. This is in conjunction with (6) above.  Side effect of speeding up compile a bit.

8) Add ONLY clauses to phys/module_physics_init.F to get rid of "call your service rep" messages.

10) Thomas Nehrkorn's updates




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4053 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d